Add files via upload
Project Description: Student Placement Prediction using Logistic Regression

Introduction:
The Student Placement Prediction project aims to develop a machine learning model using logistic regression to predict whether a student will be placed or not based on certain features. The model is trained on a custom dataset and implemented using Python. The project includes code files, a custom dataset, and a pre-trained model stored as a .pkl file.

Objective:
The primary objective of this project is to provide a predictive model that can assist educational institutions, students, and recruiters in understanding the factors influencing student placement and predicting placement outcomes accurately.